Updated folder structure and made the sidebar closing button on the sidebar in the mobile screen
