--- src/app/components/header.tsx
@@ -1,12 +1,0 @@
-import { FC } from "react";
-
-const Header: FC = () => {
-  return (
-    <header className="  mt-10 mb-10  text-center py-6  ">
-      <h1 className="text-3xl font-bold text-gray-900">PDF to Text Converter</h1>
-      <p className="text-gray-600 mt-2">Get results instantly from your PDF documents</p>
-    </header>
-  );
-};
-
-export default Header;
