+++ src/app/layout.tsx
@@ -1,8 +1,7 @@
+import LayoutWrapper from "@/components/LayoutWrapper";
 import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
 import "./globals.css";
-import LayoutWrapper from "@/app/components/LayoutWrapper";
-
 
 const geistSans = Geist({
     variable: "--font-geist-sans",
