+++ src/app/page.tsx
@@ -1,20 +1,17 @@
-import Header from "@/app/components/header";
-import UploadSection from "@/app/components/upload-section";
+import Header from "@/components/header";
+import UploadSection from "@/components/upload-section";
 
 export default function Home() {
-  return (
-    <>
-      <div>
-        <h1 className="text-7xl font-black text-center text-violet-600 pt-24">
-          Home Page
-        </h1>
-        <div className="">
-          <Header></Header>
-        </div>
-        <div className="flex justify-center mt-10 mb-10">
-        <UploadSection />
-        </div>
-      </div>
-    </>
-  );
+    return (
+        <>
+            <div className="bg-gradient-to-r from-blue-50 to-indigo-50 h-screen">
+                <div>
+                    <Header />
+                </div>
+                <div className="flex justify-center mt-10 mb-10">
+                    <UploadSection />
+                </div>
+            </div>
+        </>
+    );
 }
