+++ src/components/Sidebar.tsx
@@ -1,0 +1,166 @@
+import { Button } from "@/components/ui/button";
+import Link from "next/link";
+import React from "react";
+
+interface SidebarProps {
+    onNewChat?: () => void;
+    onNewFolder?: () => void;
+    isOpen?: boolean;
+    onClose?: () => void;
+}
+
+const Sidebar: React.FC<SidebarProps> = ({
+    onNewChat,
+    onNewFolder,
+    isOpen = false,
+    onClose,
+}) => {
+    return (
+        <>
+            {/* Mobile Overlay */}
+            {isOpen && (
+                <div
+                    className="fixed inset-0 bg-opacity-50 z-40 lg:hidden"
+                    onClick={onClose}
+                />
+            )}
+
+            {/* Sidebar */}
+            <div
+                className={`
+                fixed lg:static inset-y-0 left-0 z-50
+                w-64 h-screen bg-[#1C1C1C] text-white flex flex-col
+                transform transition-transform duration-300 ease-in-out
+                lg:transform-none font-sans
+                ${
+                    isOpen
+                        ? "translate-x-0"
+                        : "-translate-x-full lg:translate-x-0"
+                }
+            `}
+            >
+                {/* Logo */}
+                <div className="p-4 flex justify-between items-center">
+                    <Link href="/" className="flex items-center space-x-2">
+                        <span className="text-2xl font-bold tracking-tight">
+                            Talk2PDF
+                        </span>
+                    </Link>
+                    <button
+                        onClick={onClose}
+                        className="lg:hidden p-2 rounded-lg bg-gray-800 text-white hover:bg-gray-700 transition-colors"
+                        aria-label="Close Sidebar"
+                    >
+                        <svg
+                            xmlns="http://www.w3.org/2000/svg"
+                            className="h-6 w-6"
+                            fill="none"
+                            viewBox="0 0 24 24"
+                            stroke="currentColor"
+                        >
+                            <path
+                                strokeLinecap="round"
+                                strokeLinejoin="round"
+                                strokeWidth={2}
+                                d="M6 18L18 6M6 6l12 12"
+                            />
+                        </svg>
+                    </button>
+                </div>
+
+                {/* Action Buttons */}
+                <div className="px-3 space-y-2">
+                    <button
+                        onClick={onNewChat}
+                        className="w-full flex items-center justify-center space-x-2 py-2 px-4 bg-white/10 hover:bg-white/20 rounded-lg transition-colors text-base"
+                    >
+                        <svg
+                            xmlns="http://www.w3.org/2000/svg"
+                            className="h-5 w-5"
+                            viewBox="0 0 20 20"
+                            fill="currentColor"
+                        >
+                            <path
+                                fillRule="evenodd"
+                                d="M10 3a1 1 0 011 1v5h5a1 1 0 110 2h-5v5a1 1 0 11-2 0v-5H4a1 1 0 110-2h5V4a1 1 0 011-1z"
+                                clipRule="evenodd"
+                            />
+                        </svg>
+                        <span>New Chat</span>
+                    </button>
+
+                    <button
+                        onClick={onNewFolder}
+                        className="w-full flex items-center justify-center space-x-2 py-2 px-4 bg-white/10 hover:bg-white/20 rounded-lg transition-colors text-base"
+                    >
+                        <svg
+                            xmlns="http://www.w3.org/2000/svg"
+                            className="h-5 w-5"
+                            viewBox="0 0 20 20"
+                            fill="currentColor"
+                        >
+                            <path
+                                fillRule="evenodd"
+                                d="M4 4a2 2 0 012-2h4.586A2 2 0 0112 2.586L15.414 6A2 2 0 0116 7.414V16a2 2 0 01-2 2H6a2 2 0 01-2-2V4zm2 6a1 1 0 011-1h6a1 1 0 110 2H7a1 1 0 01-1-1zm1 3a1 1 0 100 2h6a1 1 0 100-2H7z"
+                                clipRule="evenodd"
+                            />
+                        </svg>
+                        <span>New Folder</span>
+                    </button>
+                </div>
+
+                {/* Chat History Section */}
+                <div className="flex-1 overflow-y-auto my-4">
+                    <div className="text-center space-y-4">
+                        <h1 className="text-xl font-semibold text-violet-500">
+                            Sign in for free to save your chat history
+                        </h1>
+                        <Button className="px-8 py-3 text-lg bg-violet-500 hover:bg-violet-600 text-white transition-colors">
+                            Sign Up
+                        </Button>
+                    </div>
+                </div>
+
+                {/* Bottom Section */}
+                <div className="p-4 border-t border-white/10">
+                    <div className="flex flex-col space-y-4">
+                        {/* Language Selector */}
+                        <div className="flex items-center space-x-2 text-sm">
+                            <svg
+                                xmlns="http://www.w3.org/2000/svg"
+                                className="h-5 w-5"
+                                viewBox="0 0 20 20"
+                                fill="currentColor"
+                            >
+                                <path
+                                    fillRule="evenodd"
+                                    d="M4.083 9h1.946c.089-1.546.383-2.97.837-4.118A6.004 6.004 0 004.083 9zM10 2a8 8 0 100 16 8 8 0 000-16zm0 2c-.076 0-.232.032-.465.262-.238.234-.497.623-.737 1.182-.389.907-.673 2.142-.766 3.556h3.936c-.093-1.414-.377-2.649-.766-3.556-.24-.56-.5-.948-.737-1.182C10.232 4.032 10.076 4 10 4zm3.971 5c-.089-1.546-.383-2.97-.837-4.118A6.004 6.004 0 0115.917 9h-1.946zm-2.003 2H8.032c.093 1.414.377 2.649.766 3.556.24.56.5.948.737 1.182.233.23.389.262.465.262.076 0 .232-.032.465-.262.238-.234.498-.623.737-1.182.389-.907.673-2.142.766-3.556zm1.166 4.118c.454-1.147.748-2.572.837-4.118h1.946a6.004 6.004 0 01-2.783 4.118zm-6.268 0C6.412 13.97 6.118 12.546 6.03 11H4.083a6.004 6.004 0 002.783 4.118z"
+                                    clipRule="evenodd"
+                                />
+                            </svg>
+                            <span>EN</span>
+                        </div>
+
+                        {/* AI Scholar Link */}
+                        <Link
+                            href="/ai-scholar"
+                            className="flex items-center space-x-2 text-sm hover:text-purple-400 transition-colors"
+                        >
+                            <svg
+                                xmlns="http://www.w3.org/2000/svg"
+                                className="h-5 w-5"
+                                viewBox="0 0 20 20"
+                                fill="currentColor"
+                            >
+                                <path d="M10.394 2.08a1 1 0 00-.788 0l-7 3a1 1 0 000 1.84L5.25 8.051a.999.999 0 01.356-.257l4-1.714a1 1 0 11.788 1.838L7.667 9.088l1.94.831a1 1 0 00.787 0l7-3a1 1 0 000-1.838l-7-3zM3.31 9.397L5 10.12v4.102a8.969 8.969 0 00-1.05-.174 1 1 0 01-.89-.89 11.115 11.115 0 01.25-3.762zM9.3 16.573A9.026 9.026 0 007 14.935v-3.957l1.818.78a3 3 0 002.364 0l5.508-2.361a11.026 11.026 0 01.25 3.762 1 1 0 01-.89.89 8.968 8.968 0 00-5.35 2.524 1 1 0 01-1.4 0zM6 18a1 1 0 001-1v-2.065a8.935 8.935 0 00-2-.712V17a1 1 0 001 1z" />
+                            </svg>
+                            <span>AI Scholar</span>
+                        </Link>
+                    </div>
+                </div>
+            </div>
+        </>
+    );
+};
+
+export default Sidebar;
