+++ tailwind.config.ts
@@ -1,0 +1,30 @@
+import type { Config } from "tailwindcss";
+
+const config: Config = {
+    content: [
+        "./src/pages/**/*.{js,ts,jsx,tsx,mdx}",
+        "./src/components/**/*.{js,ts,jsx,tsx,mdx}",
+        "./src/app/**/*.{js,ts,jsx,tsx,mdx}",
+    ],
+    theme: {
+        extend: {
+            fontFamily: {
+                sans: ["var(--font-geist-sans)"],
+                mono: ["var(--font-geist-mono)"],
+            },
+            fontSize: {
+                xs: ["0.75rem", { lineHeight: "1rem" }],
+                sm: ["0.875rem", { lineHeight: "1.25rem" }],
+                base: ["1rem", { lineHeight: "1.5rem" }],
+                lg: ["1.125rem", { lineHeight: "1.75rem" }],
+                xl: ["1.25rem", { lineHeight: "1.75rem" }],
+                "2xl": ["1.5rem", { lineHeight: "2rem" }],
+                "3xl": ["1.875rem", { lineHeight: "2.25rem" }],
+                "4xl": ["2.25rem", { lineHeight: "2.5rem" }],
+            },
+        },
+    },
+    plugins: [],
+};
+
+export default config;
